style: 메시지 컴포넌트로 분리
